final commit

+++ controllers/contactController.js
@@ -33,6 +33,11 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error('Contact not found')
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have the permission to update other user contacts")
+    }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json({success:true, message:"Contact Updated", updatedContact })
 })
@@ -43,6 +48,11 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error('Contact not found')
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403)
+        throw new Error("User don't have the permission to delete other user contacts")
+    }
+    
     await Contact.findByIdAndDelete(req.params.id)
     res.status(204).json({success:true, msg: "Contact deleted",  })
 })
